better modularization of code
