+++ src/models/scan.rs
@@ -1,0 +1,91 @@
+use std::io::Result;
+
+pub struct JsonScan<I> {
+    iter: I,
+    prev: Option<u8>,
+    state: ScanState,
+    /// (InQuotes, OutQuotes)
+    offsets: (usize, usize),
+}
+
+impl<I> JsonScan<I>
+where
+    I: Iterator<Item = Result<u8>>,
+{
+    pub fn new(iter: I) -> JsonScan<I> {
+        JsonScan {
+            iter,
+            prev: None,
+            state: ScanState::OutQuotes,
+            offsets: (0, 0),
+        }
+    }
+
+    pub fn outside_quotes(&self) -> bool {
+        match self.state {
+            ScanState::OutQuotes => true,
+            ScanState::InQuotes => false,
+        }
+    }
+
+    pub fn offsets(&self) -> (usize, usize) {
+        self.offsets
+    }
+
+    fn handle_state(&mut self) {
+        match self.prev {
+            Some(b'\\') => (),
+            _ => match self.state {
+                ScanState::InQuotes => {
+                    self.offsets.1 = 0; // Reset OutQuotes counter
+                    self.state = ScanState::OutQuotes
+                }
+                ScanState::OutQuotes => {
+                    self.offsets.0 = 0; // Reset InQuotes counter
+                    self.state = ScanState::InQuotes
+                }
+            },
+        }
+    }
+
+    fn increment_offset(&mut self) {
+        match self.state {
+            ScanState::InQuotes => self.offsets.0 += 1,
+            ScanState::OutQuotes => self.offsets.1 += 1,
+        }
+    }
+}
+
+impl<I> Iterator for JsonScan<I>
+where
+    I: Iterator<Item = Result<u8>>,
+{
+    type Item = Result<u8>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        match self.iter.next() {
+            Some(Ok(b @ b'"')) => {
+                self.handle_state();
+                self.prev = Some(b);
+                Some(Ok(b))
+            }
+            Some(Ok(b)) => {
+                self.increment_offset();
+                self.prev = Some(b);
+                Some(Ok(b))
+            }
+            Some(Err(e)) => {
+                self.increment_offset();
+                self.prev = None;
+                Some(Err(e))
+            }
+            None => None,
+        }
+    }
+}
+
+#[derive(Clone, Copy, PartialEq)]
+enum ScanState {
+    InQuotes,
+    OutQuotes,
+}
